isPrime
For input integer n, return True if it’s a prime number and False if
it’s not.

I had two attempts at this challenge.
Key point is that you only need to evaluate n%1 up to the square root
of n. This is because if n is not a prime number there will be two
numbers that multiply together to make n and the smaller of the two
will always be less than the square root of n.